Simplify types tests
